move ConsoleControl's AnsiColorWriter to separate file
(for easier comparison with other copies of that code)
